Add files via upload
Time series analysis and dynamic modeling are important to study topics with a variety of applications in business, economics, finance, and computer science. The purpose of time series analysis is to examine the projected path of time series observations, build a model to characterize the data structure, and then predict the time series’ future values. In economics, Real GDP Per Capita serves to reflect whether the nation is progressing or declining in economic growth. An increase in Real GDP signals that the economy is doing well, given that products and services are growing in value, while a decrease signals the opposite. Forecasting of GDP of production industries plays a major role in the optimal decision formulas for government and industry in the United States (US). In this project, we will forecast whether the country’s economy is expanding or contracting due to its output on the factors of production. We can use Quandl to obtain data going back to 1947 for the Real gross domestic product per capita, Federal Reserve Bank of St. Louis uses the symbol “rxVQhZ8_nxxeo2yy4Uz4” (Quandle code : FRED/GDPC1). We will use historical values of GDP per capita for the US, from January 01, 1947 to July 01, 2021, for all four quarters per year. The GDP data has been seasonally adjusted. The quarters were collected (299 observations). The first 267 are to be used as training data, while the others are to be used as a test set. A train and test set was created. The range was as follows:

Training Set Range: 01-01-1947 - 01-10-2013

Test Set Range: 01-01-2014 - 01-07-2021